+++ refactor-me.Tests/Integration/IntegrationTestBase.cs
@@ -1,37 +1,48 @@
 ﻿using System.Data;
 using System.Data.SqlClient;
 using System.IO;
+using System.Transactions;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace refactor_me.Tests.Integration
 {
+    /// <summary>
+    /// Base class used for integration test.
+    /// </summary>
+    [TestClass]
     public abstract class IntegrationTestBase
     {
+        private TransactionScope _scope;
+
         private const string TestConnectionString =
-                @"Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename={DataDirectory}\TestDatabase.mdf;Integrated Security=True";
+                @"Data Source=(LocalDB)\MSSQLLocalDB;AttachDbFilename={DataDirectory}\Database.mdf;Integrated Security=True";
 
         protected IDbConnection GetTestDbConnection()
         {
             return new SqlConnection(TestConnectionString.Replace("{DataDirectory}", GetDbPath()));
         }
 
-        public static void SetUp()
+        public void BeginTransaction()
         {
-            var dbPath = GetDbPath();
+            _scope = new TransactionScope();
 
-            try  // Super hack to avoid "file already being used..." exceptions when running multiple test fixtures
-            {
-                if (File.Exists(Path.Combine(dbPath, "TestDatabase.mdf")))
-                {
-                    File.Delete(Path.Combine(dbPath, "TestDatabase.mdf"));
-                    File.Delete(Path.Combine(dbPath, "TestDatabase_log.ldf"));
-                }
+        }
 
-                File.Copy(Path.Combine(dbPath, "Database.mdf"), Path.Combine(dbPath, "TestDatabase.mdf"));
-            }
-            catch
-            {
-                // ignored
-            }
+        public void RollbackTransaction()
+        {
+            _scope?.Dispose();
+        }
+
+        [TestInitialize]
+        public void Initialize()
+        {
+            BeginTransaction();
+        }
+
+        [TestCleanup]
+        public void Cleanup()
+        {
+            RollbackTransaction();
         }
 
         private static string GetDbPath()
