+++ refactor-me/Models/Product.cs
@@ -1,0 +1,23 @@
+﻿using System;
+
+namespace refactor_me.Models
+{
+    public class Product
+    {
+        public Guid Id { get; set; }
+
+        public string Name { get; set; }
+
+        public string Description { get; set; }
+
+        public decimal Price { get; set; }
+
+        public decimal DeliveryPrice { get; set; } 
+   
+
+        public Product()
+        {
+            Id = Guid.NewGuid(); 
+        }
+    }
+}
