+++ refactor-me/Models/ProductOption.cs
@@ -1,0 +1,20 @@
+﻿using System;
+
+namespace refactor_me.Models
+{
+    public class ProductOption
+    {
+        public Guid Id { get; set; }
+
+        public Guid ProductId { get; set; }
+
+        public string Name { get; set; }
+
+        public string Description { get; set; }  
+
+        public ProductOption()
+        {
+            Id = Guid.NewGuid();
+        } 
+    }
+}
