+++ refactor-me/Models/ProductOptions.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace refactor_me.Models
+{
+    public class ProductOptions
+    {
+        public IEnumerable<ProductOption> Items { get; set; } 
+    }
+}
