+++ refactor-me/Models/Products.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data.SqlClient;
-using Newtonsoft.Json;
+﻿using System.Collections.Generic;
 
 namespace refactor_me.Models
 {
@@ -9,86 +6,5 @@ namespace refactor_me.Models
     {
         public IEnumerable<Product> Items { get; set; }  
    
-    }
-
-    public class Product
-    {
-        public Guid Id { get; set; }
-
-        public string Name { get; set; }
-
-        public string Description { get; set; }
-
-        public decimal Price { get; set; }
-
-        public decimal DeliveryPrice { get; set; } 
-   
-
-        public Product()
-        {
-            Id = Guid.NewGuid(); 
-        }
-    }
-
-    public class ProductOptions
-    {
-        public IEnumerable<ProductOption> Items { get; set; } 
-    }
-
-    public class ProductOption
-    {
-        public Guid Id { get; set; }
-
-        public Guid ProductId { get; set; }
-
-        public string Name { get; set; }
-
-        public string Description { get; set; }
-
-        [JsonIgnore]
-        public bool IsNew { get; }
-
-        public ProductOption()
-        {
-            Id = Guid.NewGuid();
-            IsNew = true;
-        }
-
-        public ProductOption(Guid id)
-        {
-            IsNew = true;
-            var conn = DbHelpers.GetConnection();
-            var cmd = new SqlCommand($"select * from productoption where id = '{id}'", conn);
-            conn.Open();
-
-            var rdr = cmd.ExecuteReader();
-            if (!rdr.Read())
-                return;
-
-            IsNew = false;
-            Id = Guid.Parse(rdr["Id"].ToString());
-            ProductId = Guid.Parse(rdr["ProductId"].ToString());
-            Name = rdr["Name"].ToString();
-            Description = (DBNull.Value == rdr["Description"]) ? null : rdr["Description"].ToString();
-        }
-
-        public void Save()
-        {
-            var conn = DbHelpers.GetConnection();
-            var cmd = IsNew ?
-                new SqlCommand($"insert into productoption (id, productid, name, description) values ('{Id}', '{ProductId}', '{Name}', '{Description}')", conn) :
-                new SqlCommand($"update productoption set name = '{Name}', description = '{Description}' where id = '{Id}'", conn);
-
-            conn.Open();
-            cmd.ExecuteNonQuery();
-        }
-
-        public void Delete()
-        {
-            var conn = DbHelpers.GetConnection();
-            conn.Open();
-            var cmd = new SqlCommand($"delete from productoption where id = '{Id}'", conn);
-            cmd.ExecuteReader();
-        }
     }
 }
