+++ refactor-me/refactor-me.csproj
@@ -102,6 +102,9 @@
     </Compile>
     <Compile Include="Infrastructure\PoorManControllerActivator.cs" />
     <Compile Include="Models\DbHelpers.cs" />
+    <Compile Include="Models\Product.cs" />
+    <Compile Include="Models\ProductOption.cs" />
+    <Compile Include="Models\ProductOptions.cs" />
     <Compile Include="Models\Products.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Repositories\ProductRepository.cs" />
